# openapi/yl-api.yaml
openapi: 3.0.3
info:
  title: 银龄守护 · 统一 API（yl-server）
  description: |
    四端（微信小程序 / 抖音小程序 / iOS / 安卓）共用同一套 API，由后端统一供给
    （PRD YL-PRD-v1.0 §8「统一原则」）。

    ## 统一约定
    - **版本**：所有业务接口前缀 `/api/v1`，版本升级以新增路径而非破坏性修改实现。
    - **鉴权**：`Authorization: Bearer <accessToken>`；access 2h / refresh 14d（JWT 双令牌）。
      四端登录方式差异（微信/抖音授权、Apple ID/Face ID、账号密码）由 `/api/v1/auth/*` 统一换取 yl 自有令牌。
    - **统一响应**：HTTP 状态码只表达传输层语义；业务语义一律走响应体 `code`。
      `code=0` 成功，非 0 见「错误码字典」。
    - **幂等**：写操作（提交评估、生成报告、上报、导出）必须携带 `Idempotency-Key`（客户端 UUID）；
      服务端按 key 去重，重复请求直接返回首次结果，不重复执行。
    - **分页**：`pageNum`（从 1 起）/ `pageSize`（1-100，默认 20）。
    - **脱敏**：手机号/身份证/姓名等敏感字段默认脱敏返回（如 `138****5678`），
      需查看明文须走 `/api/v1/security/reveal`（二次验证 + 留痕）。
    - **追踪**：响应体与响应头 `X-Trace-Id` 均返回 traceId，便于全链路排障。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发（docker-compose 一键启动）
  - url: https://api.yl.example.com
    description: 生产（示例域名）
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 统一鉴权令牌（access token）
...
  schemas:
    R:
      type: object
      description: 统一响应包裹
      properties:
        code: { type: integer, description: 0=成功；非0=错误码 }
        message: { type: string }
        data: { nullable: true }
        traceId: { type: string }
        timestamp: { type: integer, format: int64 }
      required: [code, message, traceId]
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: 秒 }
        userId: { type: integer, format: int64 }
        roles: { type: array, items: { type: string } }
        orgId: { type: integer, format: int64, nullable: true }
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: 刷新令牌
      x-required-permission: none
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
              required: [refreshToken]
      responses:
        '200':
          description: 新的令牌对
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - properties:
                      data: { $ref: '#/components/schemas/TokenPair' }
  /api/v1/auth/second-verify:
    post:
      tags: [auth]
      summary: 换取二次验证一次性凭据（敏感操作前置）
      description: |
        敏感操作（导出 / 作废评估单 / 解绑亲情 / 查看明文）前调用。校验通过后返回一次性凭据，
        后续业务请求以 `X-Second-Verify-Token` 头携带；凭据一次性消费、300 秒有效（Redis GETDEL）。
      x-required-permission: none
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [SCAN, PHONE, FACE]
                  description: 二次验证方式
                credential:
                  type: string
                  description: 渠道凭据（联调期为配置口令；生产为短信/扫码/人脸渠道结果）
              required: [method, credential]
      responses:
        '200':
          description: 一次性凭据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - properties:
                      data:
                        type: object
                        properties:
                          token: { type: string }
                          expiresIn:
                            type: integer
                            description: 有效期（秒），当前为 300
...